Elemental system formula update. Also using byte data type for elements.
git-svn-id: http://svn.l2jserver.com/trunk@3190 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/L2ItemInstance.java
@@ -947,7 +947,7 @@ public final class L2ItemInstance extends L2Object
 		return _elementals;
 	}
 
-	public int getAttackElementType()
+	public byte getAttackElementType()
 	{
 		if (isWeapon() && _elementals != null)
 			return _elementals.getElement();
